+++ packages/nuxt/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @clerk/nuxt
 
+## 1.13.27
+
+### Patch Changes
+
+- Updated dependencies [[`e5f213f`](https://github.com/clerk/javascript/commit/e5f213f2ab2773992f7156fca760f836ff19d505)]:
+  - @clerk/shared@3.47.3
+  - @clerk/backend@2.33.1
+  - @clerk/types@4.101.21
+  - @clerk/vue@1.17.19
+
 ## 1.13.26
 
 ### Patch Changes